JAMES-2904 Introduce a KeyspaceFactory
Handles the responsibility of creating the keyspace, if it does not
already exists.
