+++ backends-common/cassandra/src/main/java/org/apache/james/backends/cassandra/init/ResilientClusterProvider.java
@@ -62,7 +62,7 @@ public class ResilientClusterProvider implements Provider<Cluster> {
         Cluster cluster = ClusterFactory.create(configuration);
         return () -> {
             try {
-                ClusterFactory.createKeyspace(configuration, cluster);
+                KeyspaceFactory.createKeyspace(configuration, cluster);
                 return cluster;
             } catch (Exception e) {
                 cluster.close();
